requirements.txt
praw==7.1.0
python-telegram-bot==13.1
PyYAML==5.3.1
boto3==1.16.49